src/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tqdm
scikit-learn
seaborn
matplotlib
Pillow
argparse
